WIP on refactoring components
