+++ front/components/ideaDetail/IdeaCard.vue
@@ -1,92 +1,111 @@
 <template>
-  <v-row
-    ref="page"
-    align="stretch"
-    :class="{ 'fixed-height': !isExpanded && !preview }"
-    class="elevation-2 ma-1 card"
-    :style="additionalStyling"
+  <card
+    :class="{ 'fixed-height': !isExpanded }"
+    :additional-styling="additionalStyling"
+    @click="expandToggle"
   >
-    <div class="tap-bar" @click="expandToggle"></div>
-    <slot :expand-toggle="expandToggle"></slot>
-  </v-row>
+    <v-icon v-if="closeBtn" class="close-btn" @click="$emit('exit-pressed')"
+      >mdi mdi-close</v-icon
+    >
+    <idea-edit
+      v-if="editMode"
+      :idea="idea"
+      :idea-tags="ideaTags"
+      @updated="onUpdate"
+      @cancel="editMode = false"
+    />
+    <idea-show
+      v-else
+      :idea="idea"
+      :expanded="isExpanded"
+      :preview="preview"
+      :idea-tags="ideaTags"
+      @edit="editMode = true"
+      @close="isExpanded = false"
+      @view-preview="$emit('view-preview')"
+      @updated="onUpdate"
+    />
+  </card>
 </template>
 
 <script>
+import merge from 'lodash/merge'
+import Card from '@/components/shared/Card'
+import IdeaShow from '@/components/ideaDetail/IdeaShow'
+import IdeaEdit from '@/components/ideaDetail/IdeaEdit'
+import getIdeaTags from '@/graphql/query/getIdeaTags'
+
 export default {
-  name: 'IdeaCard',
+  components: { IdeaEdit, IdeaShow, Card },
+
   props: {
-    allowMobileScroll: Boolean,
-    expanded: Boolean,
     preview: Boolean,
-    additionalStyling: {
-      type: Object,
-      default: Object
-    }
+    idea: { type: Object, required: true },
+    additionalStyling: { type: Object, default: Object },
+    expanded: { type: Boolean, default: false },
+    closeBtn: { type: Boolean, default: false }
   },
+
   data() {
     return {
-      isExpanded: this.expanded === undefined ? false : this.expanded
+      ideaTags: [],
+      isExpanded: false,
+      editMode: false
     }
   },
+
   mounted() {
-    // Disable vertical scroll if card is not expanded
-    if (!this.allowMobileScroll) {
-      this.$refs.page.addEventListener(
-        'touchmove',
-        this.preventScrollOnMobileWhenCardIsNotExpanded
-      )
+    this.loadIdeaTags()
+    if (this.$route.query.comment) {
+      setTimeout(() => this.commentsBtnClick(), 500)
     }
+    this.isExpanded = this.isMobile ? this.expanded : true
   },
+
   methods: {
-    preventScrollOnMobileWhenCardIsNotExpanded(event) {
-      if (this.isExpanded === false) {
-        event.preventDefault()
-      } else {
-        return true
-      }
-    },
     expandToggle() {
+      if (this.expanded || !this.isMobile) {
+        return
+      }
       this.isExpanded = !this.isExpanded
-      this.$emit('expand-toggle')
+      this.$emit('expand-toggle', this.isExpanded)
+    },
+
+    async loadIdeaTags() {
+      const ideaTags = []
+      if (this.$store.getters['cognito/isLoggedIn']) {
+        const tag = await this.$amplifyApi.graphql({
+          query: getIdeaTags,
+          variables: { ideaId: this.idea.ideaId }
+        })
+
+        for (let i = 0; i < tag.data.ideaTags.length; i++) {
+          ideaTags.push(tag.data.ideaTags[i].tag)
+        }
+      }
+      this.ideaTags = ideaTags
+    },
+
+    onUpdate(idea) {
+      this.$emit('update', merge({}, this.idea, idea))
+      this.editMode = false
     }
   }
 }
 </script>
 
-<style lang="scss" scoped>
-.card {
-  position: relative;
-  padding: 1rem;
-  background-color: white;
-  border-radius: 8px;
-  border: 1px solid $light-grey !important;
-  box-shadow: $card-shadow !important;
-  overflow: hidden;
+<style la lang="scss" scoped>
+.close-btn {
+  position: absolute;
+  z-index: 1001;
+  top: 10px;
+  right: 10px;
+  color: $primary-color;
+  font-size: 20px;
 
-  @media only screen and (min-width: $screen-md-min) {
-    margin: 0 auto !important;
-    margin-top: 2vh !important;
-
-    .fixed-height {
-      min-height: 80vh;
-    }
-  }
-
-  /* .rounded doesn't work because i'm applying this to a .row (which i shouldn't) */
-}
-
-@media only screen and (max-width: $screen-sm-max) {
-  .fixed-height {
-    overflow-y: hidden;
-  }
-
-  .tap-bar {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 10%;
-    z-index: 100;
+  @media (max-width: $screen-sm-max) {
+    left: 10px;
+    justify-content: start;
   }
 }
 </style>
