+++ front/components/ideaDetail/SwipeExplainer.vue
@@ -1,11 +1,6 @@
 <template>
-  <idea-card
-    :allow-mobile-scroll="allowMobileScroll"
-    :preview="preview"
-    :additional-styling="additionalStyling"
-    @expand-toggle="isExpanded = !isExpanded"
-  >
-    <v-col cols="12" :md="preview ? '' : 8" class="idea-part">
+  <card :additional-styling="additionalStyling">
+    <v-col cols="12" class="idea-part">
       <v-row class="idea-part__header no-gutters" no-gutters>
         <v-col class="idea-part__header__title">
           <h2 class="idea-part__header__title__label">
@@ -32,42 +27,19 @@
         </div>
       </div>
     </v-col>
-  </idea-card>
+  </card>
 </template>
 
 <script>
-import IdeaCard from '@/components/ideaDetail/IdeaCard'
+import Card from '@/components/shared/Card'
 
 export default {
   components: {
-    IdeaCard
+    Card
   },
 
   props: {
-    additionalStyling: {
-      type: Object,
-      default: Object
-    },
-
-    allowMobileScroll: {
-      type: Boolean,
-      default: false
-    },
-
-    closeBtn: {
-      type: Boolean,
-      default: false
-    },
-
-    preview: {
-      type: Boolean,
-      default: Boolean
-    },
-
-    expanded: {
-      type: Boolean,
-      default: false
-    }
+    additionalStyling: { type: Object, default: Object }
   },
 
   data() {
@@ -100,62 +72,8 @@ export default {
 </script>
 
 <style scoped lang="scss">
-.comments-section {
-  display: flex;
-  align-items: flex-end;
-}
-
-.pointer {
-  cursor: pointer;
-}
-
-.close-btn {
-  position: absolute;
-  z-index: 1001;
-  top: 10px;
-  right: 10px;
-  color: $primary-color;
-  font-size: 20px;
-
-  @media (max-width: $screen-sm-max) {
-    left: 10px;
-    justify-content: start;
-  }
-}
-
-.idea-name-field {
-  font-size: 24px;
-}
-
-.idea-date {
-  margin-top: 2px;
-}
-
-.idea-link {
-  text-decoration: none;
-}
-
 .idea-part {
   position: relative;
-
-  &__author-avatar {
-    width: 24px;
-    margin-bottom: 5px;
-    border-radius: 100%;
-  }
-
-  &__info {
-    &__author {
-      &__link {
-        text-transform: capitalize;
-        text-decoration: none;
-        color: $disabled-purple;
-        &:hover {
-          text-decoration: underline;
-        }
-      }
-    }
-  }
 
   &__idea-content {
     @media only screen and (min-width: $screen-md-min) {
@@ -165,50 +83,11 @@ export default {
     }
   }
 
-  &__content-container {
-    word-break: break-word;
-    margin-top: 1rem;
-    &__idea-editor {
-      ::v-deep .trix-content {
-        overflow-y: auto;
-      }
-    }
-  }
-
-  &__edit-buttons-panel {
-    text-align: right;
-    margin: 10px 0;
-    padding: 0 5px;
-  }
-
   &__header {
     &__title {
       text-transform: capitalize;
       &__label {
         font-weight: 100;
-      }
-    }
-  }
-
-  &__tags-panel {
-    /*border-top: 1px solid #ebe7ed;*/
-    /*padding-left: 10px;*/
-
-    .tagsContainer {
-      margin-top: 20px;
-
-      .tag {
-        margin: 4px;
-        background-color: white;
-        color: $disabled-purple;
-        border-radius: 5px;
-        border: 1px solid $disabled-purple;
-      }
-    }
-
-    &__tags-editor {
-      ::v-deep .v-input__icon {
-        display: none;
       }
     }
   }
