+++ front/components/shared/Card.vue
@@ -1,0 +1,63 @@
+<template>
+  <div
+    ref="card"
+    class="elevation-2 ma-1 card"
+    :style="additionalStyling"
+    @click="e => $emit('click', e)"
+  >
+    <slot></slot>
+  </div>
+</template>
+
+<script>
+export default {
+  props: {
+    lightboxMode: Boolean,
+    additionalStyling: { type: [Object, String], default: Object },
+    allowMobileScroll: { type: Boolean, default: false }
+  },
+
+  mounted() {
+    // if (!this.allowMobileScroll) {
+    //   this.$refs.card.addEventListener('touchmove', this.preventScrollOnMobile)
+    // }
+  },
+
+  methods: {
+    preventScrollOnMobile(event) {
+      if (this.isExpanded === false) {
+        event.preventDefault()
+      } else {
+        return true
+      }
+    }
+  }
+}
+</script>
+
+<style lang="scss" scoped>
+.card {
+  position: relative;
+  padding: 1rem;
+  background-color: white;
+  border-radius: 8px;
+  border: 1px solid $light-grey !important;
+  box-shadow: $card-shadow !important;
+  overflow: hidden;
+
+  @media only screen and (min-width: $screen-md-min) {
+    margin: 0 auto !important;
+    margin-top: 2vh !important;
+
+    .fixed-height {
+      min-height: 80vh;
+    }
+  }
+}
+
+@media only screen and (max-width: $screen-sm-max) {
+  .fixed-height {
+    overflow-y: hidden;
+  }
+}
+</style>
