+++ front/pages/i/_shortId/_slug.vue
@@ -3,6 +3,7 @@
     <div
       class="light-box-bg"
       :class="{ 'light-box-expanded': isExpanded }"
+      @click="$refs.page.expandToggle()"
     ></div>
     <layout :hide-mobile-nav="isExpanded" :hide-slide-menu="hideSlideMenu">
       <template v-slot:header>
@@ -23,7 +24,7 @@
         @animation-out-end="animationOutEnd"
       >
         <template v-slot:background>
-          <idea-card-skeleton />
+          <idea-card-skeleton :additional-styling="{ 'min-height': '71vh' }" />
         </template>
         <template v-slot="{ rotationStyle }">
           <swipe-explainer
@@ -32,15 +33,16 @@
             :style="rotationStyle"
             :additional-styling="{ 'min-height': '71vh' }"
           ></swipe-explainer>
-          <full-idea
+          <idea-card
             v-else
             ref="page"
             class="card"
             :idea="idea"
             :style="rotationStyle"
             :additional-styling="{ 'min-height': '71vh' }"
-            @toggle-expand="isExpanded = !isExpanded"
-          ></full-idea>
+            @update="i => (idea = i)"
+            @expand-toggle="val => (isExpanded = val)"
+          ></idea-card>
         </template>
       </swiper>
     </layout>
@@ -52,7 +54,6 @@ import clip from 'text-clipper'
 import Cookies from 'js-cookie'
 import { mapGetters } from 'vuex'
 import Layout from '@/components/layout/Layout'
-import FullIdea from '@/components/ideaDetail/FullIdea.vue'
 import {
   getNewIdeas,
   getTopIdeas
@@ -63,12 +64,13 @@ import incrementIdeaViews from '@/graphql/mutations/incrementIdeaViews'
 import IdeaCardSkeleton from '@/components/ideaDetail/IdeaCardSkeleton'
 import SwipeExplainer from '@/components/ideaDetail/SwipeExplainer'
 import CategoriesSubHeader from '@/components/layout/CategoriesSubHeader'
+import IdeaCard from '@/components/ideaDetail/IdeaCard'
 
 export default {
   components: {
+    IdeaCard,
     Layout,
     Swiper,
-    FullIdea,
     IdeaCardSkeleton,
     CategoriesSubHeader,
     SwipeExplainer
@@ -106,7 +108,8 @@ export default {
       idea: null,
       expandedState: false,
       showExplainer: false,
-      category: 'top'
+      category: 'top',
+      isExpanded: false
     }
   },
 
@@ -114,16 +117,7 @@ export default {
     ...mapGetters({
       userId: 'userData/userId',
       isAuthenticated: 'userData/isAuthenticated'
-    }),
-
-    isExpanded: {
-      set() {
-        this.expandedState = !this.expandedState
-      },
-      get() {
-        return this.expandedState === undefined ? false : this.expandedState
-      }
-    }
+    })
   },
 
   created() {
